rewrite banner rendering
